INTERNAL: Fix adding and removing row in reordered panels.

+++ src/model/entities/FormScreenLifecycle/FormScreenLifecycle.tsx
@@ -45,19 +45,20 @@ import { getUserFilterLookups } from "../../selectors/DataView/getUserFilterLook
 import _ from "lodash";
 import { ChangeMasterRecordDialog } from "../../../gui/Components/Dialogs/ChangeMasterRecordDialog";
 import { getFormScreenLifecycle } from "../../selectors/FormScreen/getFormScreenLifecycle";
-import { selectFirstRow} from "../../actions/DataView/selectFirstRow";
+import { selectFirstRow } from "../../actions/DataView/selectFirstRow";
 import { YesNoQuestion } from "gui/Components/Dialogs/YesNoQuestion";
 import { getProperties } from "model/selectors/DataView/getProperties";
 import { getWorkbench } from "model/selectors/getWorkbench";
 import { shouldProceedToChangeRow } from "model/actions-ui/DataView/TableView/shouldProceedToChangeRow";
 import { getGroupingConfiguration } from "model/selectors/TablePanelView/getGroupingConfiguration";
-import {IDataViewToolbarUI} from "modules/DataView/DataViewUI";
-import {IFormPerspectiveDirector} from "modules/DataView/Perspective/FormPerspective/FormPerspectiveDirector";
-import {selectLastRow} from "model/actions/DataView/selectLastRow";
-import {startEditingFirstCell} from "model/actions/DataView/startEditingFirstCell";
+import { IDataViewToolbarUI } from "modules/DataView/DataViewUI";
+import { IFormPerspectiveDirector } from "modules/DataView/Perspective/FormPerspective/FormPerspectiveDirector";
+import { selectLastRow } from "model/actions/DataView/selectLastRow";
+import { startEditingFirstCell } from "model/actions/DataView/startEditingFirstCell";
 import { getTablePanelView } from "model/selectors/TablePanelView/getTablePanelView";
 import { getFocusManager } from "model/selectors/DataView/getFocusManager";
 import { wait } from "@testing-library/react";
+import { getDataSourceFieldByName } from "model/selectors/DataSources/getDataSourceFieldByName";
 
 enum IQuestionSaveDataAnswer {
   Cancel = 0,
@@ -105,8 +106,12 @@ export class FormScreenLifecycle02 implements IFormScreenLifecycle02 {
     yield* this.copyRow(entity, gridId, rowId);
   }
 
-  *onDeleteRow(entity: string, rowId: string): Generator<unknown, any, unknown> {
-    yield* this.onRequestDeleteRow(entity, rowId);
+  *onDeleteRow(
+    entity: string,
+    rowId: string,
+    dataView: IDataView
+  ): Generator<unknown, any, unknown> {
+    yield* this.onRequestDeleteRow(entity, rowId, dataView);
   }
 
   *onSaveSession(): Generator<unknown, any, unknown> {
@@ -147,9 +152,9 @@ export class FormScreenLifecycle02 implements IFormScreenLifecycle02 {
     );
   }
 
-  *onRequestDeleteRow(entity: string, rowId: string) {
+  *onRequestDeleteRow(entity: string, rowId: string, dataView: IDataView) {
     if ((yield this.questionDeleteData()) === IQuestionDeleteDataAnswer.Yes) {
-      yield* this.deleteRow(entity, rowId);
+      yield* this.deleteRow(entity, rowId, dataView);
     }
   }
 
@@ -187,8 +192,8 @@ export class FormScreenLifecycle02 implements IFormScreenLifecycle02 {
         yield* this.saveSession();
         yield* this.refreshSession();
         return;
-        case IQuestionSaveDataAnswer.NoSave:
-        if(!this.isReadData){
+      case IQuestionSaveDataAnswer.NoSave:
+        if (!this.isReadData) {
           yield* this.revertChanges();
         }
         yield* this.refreshSession();
@@ -196,8 +201,8 @@ export class FormScreenLifecycle02 implements IFormScreenLifecycle02 {
     }
   }
   *revertChanges(): Generator<unknown, any, unknown> {
-      const api = getApi(this);
-      yield api.revertChanges({sessionFormIdentifier: getSessionId(this)});
+    const api = getApi(this);
+    yield api.revertChanges({ sessionFormIdentifier: getSessionId(this) });
   }
 
   *onWorkflowNextClick(event: any): Generator {
@@ -375,19 +380,25 @@ export class FormScreenLifecycle02 implements IFormScreenLifecycle02 {
       }
     }
     yield* this.startAutorefreshIfNeeded();
+
+    console.log("----- DataViews debug -----");
+    for (let dv of getFormScreen(this).dataViews) {
+      console.log(dv.id, "-", dv.name, "-", dv.orderProperty, "-", dv.orderMember);
+    }
+    console.log("----- =============== -----");
   }
 
-  sortAndFilterReaction(dataView : IDataView){
-    const self=this;
+  sortAndFilterReaction(dataView: IDataView) {
+    const self = this;
     flow(function* () {
       if (!(yield shouldProceedToChangeRow(dataView))) {
         return;
       }
-      yield dataView.lifecycle.runRecordChangedReaction(function*() {
+      yield dataView.lifecycle.runRecordChangedReaction(function* () {
         const groupingConfig = getGroupingConfiguration(dataView);
-        if(groupingConfig.isGrouping){
+        if (groupingConfig.isGrouping) {
           dataView.serverSideGrouper.refresh();
-        }else{
+        } else {
           yield self.readFirstChunkOfRowsWithGateDebounced(dataView);
         }
       });
@@ -445,8 +456,7 @@ export class FormScreenLifecycle02 implements IFormScreenLifecycle02 {
         RowOffset: 0,
         ColumnNames: getColumnNamesToLoad(rootDataView),
       });
-    }
-    finally{
+    } finally {
       this.monitor.inFlow--;
     }
   }
@@ -478,8 +488,7 @@ export class FormScreenLifecycle02 implements IFormScreenLifecycle02 {
         MasterRowId: undefined,
         AggregatedColumns: aggregations,
       });
-    }
-    finally{
+    } finally {
       this.monitor.inFlow--;
     }
   }
@@ -510,8 +519,7 @@ export class FormScreenLifecycle02 implements IFormScreenLifecycle02 {
         MasterRowId: undefined,
         AggregatedColumns: aggregations,
       });
-    }
-    finally{
+    } finally {
       this.monitor.inFlow--;
     }
   }
@@ -601,27 +609,26 @@ export class FormScreenLifecycle02 implements IFormScreenLifecycle02 {
     }
   }
 
-  private *runUpdateObject(dataView: IDataView){
-    const updateData = dataView.dataTable.getDirtyValueRows().map(row => {
+  private *runUpdateObject(dataView: IDataView) {
+    const updateData = dataView.dataTable.getDirtyValueRows().map((row) => {
       return {
         RowId: dataView.dataTable.getRowId(row),
-        Values: map2obj(dataView.dataTable.getDirtyValues(row))
-      }})
-    if(!updateData || updateData.length === 0){
+        Values: map2obj(dataView.dataTable.getDirtyValues(row)),
+      };
+    });
+    if (!updateData || updateData.length === 0) {
       return false;
     }
     const api = getApi(this);
     const formScreen = getFormScreen(this);
     const self = this;
-    const updateObjectResult = yield* formScreen.dataUpdateCRS.runGenerator<any>(
-      function* () {
-        return yield api.updateObject({
-          SessionFormIdentifier: getSessionId(self),
-          Entity: dataView.entity,
-          UpdateData: updateData,
-        });
-      }
-    );
+    const updateObjectResult = yield* formScreen.dataUpdateCRS.runGenerator<any>(function* () {
+      return yield api.updateObject({
+        SessionFormIdentifier: getSessionId(self),
+        Entity: dataView.entity,
+        UpdateData: updateData,
+      });
+    });
     dataView.focusManager.stopAutoFocus();
     yield* processCRUDResult(dataView, updateObjectResult);
     return true;
@@ -639,10 +646,12 @@ export class FormScreenLifecycle02 implements IFormScreenLifecycle02 {
         return yield api.updateObject({
           SessionFormIdentifier: getSessionId(self),
           Entity: dataView.entity,
-          UpdateData: [{
-            RowId: dataView.dataTable.getRowId(row),
-            Values: changes
-          }]
+          UpdateData: [
+            {
+              RowId: dataView.dataTable.getRowId(row),
+              Values: changes,
+            },
+          ],
         });
       });
 
@@ -706,20 +715,20 @@ export class FormScreenLifecycle02 implements IFormScreenLifecycle02 {
     500
   );
 
-  private getNewRowValues(){
-    if(!this.userOrderedDataView) {
-      return {}
+  private getNewRowValues(targetDataView: IDataView) {
+    if (!targetDataView.orderMember) {
+      return;
     }
 
-    const orderProperty = this.userOrderedDataView.orderProperty!;
-    const orderValues = this.userOrderedDataView.tableRows
-      .filter(row => Array.isArray)
-      .map(row => (row as any[])[orderProperty.dataIndex] as number);
-    const nextOrderValue = Math.max(...orderValues) + 1
-    const orderPropName = orderProperty.name;
+    const orderMember = targetDataView.orderMember;
+    const dataSourceField = getDataSourceFieldByName(targetDataView, orderMember);
+    const orderValues = targetDataView.tableRows
+      .filter((row) => Array.isArray)
+      .map((row) => (row as any[])[dataSourceField!.index] as number);
+    const nextOrderValue = orderValues.length > 0 ? Math.max(...orderValues) + 1 : 0;
     const values = {} as any;
-    values[orderProperty.name] = nextOrderValue;
-    return values
+    values[orderMember] = nextOrderValue;
+    return values;
   }
 
   *createRow(entity: string, gridId: string) {
@@ -735,19 +744,19 @@ export class FormScreenLifecycle02 implements IFormScreenLifecycle02 {
           SessionFormIdentifier: getSessionId(this),
           Entity: entity,
           RequestingGridId: gridId,
-          Values: this.getNewRowValues(),
+          Values: this.getNewRowValues(targetDataView),
           Parameters: { ...getBindingParametersFromParent(targetDataView) },
         });
       } finally {
         formScreen.dataUpdateCRS.leave();
       }
       yield* processCRUDResult(targetDataView, createObjectResult);
-      if(targetDataView.newRecordView === "0" && targetDataView.activateFormView){
+      if (targetDataView.newRecordView === "0" && targetDataView.activateFormView) {
         yield* targetDataView.activateFormView();
-      }else{
-        if(!targetDataView.isFormViewActive()){
+      } else {
+        if (!targetDataView.isFormViewActive()) {
           yield* startEditingFirstCell(targetDataView)();
-        }else{
+        } else {
           getFocusManager(targetDataView).forceAutoFocus();
         }
       }
@@ -771,7 +780,7 @@ export class FormScreenLifecycle02 implements IFormScreenLifecycle02 {
           OriginalId: rowId,
           RequestingGridId: gridId,
           Entities: [entity],
-          ForcedValues: this.getNewRowValues(),
+          ForcedValues: this.getNewRowValues(targetDataView),
         });
       } finally {
         formScreen.dataUpdateCRS.leave();
@@ -782,11 +791,11 @@ export class FormScreenLifecycle02 implements IFormScreenLifecycle02 {
     }
   }
 
-  private get userOrderedDataView(){
-    return getFormScreen(this).dataViews.find(dataView => dataView.orderProperty);
-  }
+  /*private get userOrderedDataView() {
+    return getFormScreen(this).dataViews.find((dataView) => dataView.orderProperty);
+  }*/
 
-  *deleteRow(entity: string, rowId: string) {
+  *deleteRow(entity: string, rowId: string, targetDataView: IDataView) {
     try {
       this.monitor.inFlow++;
       const api = getApi(this);
@@ -795,9 +804,9 @@ export class FormScreenLifecycle02 implements IFormScreenLifecycle02 {
       try {
         yield* formScreen.dataUpdateCRS.enterGenerator();
 
-        if(this.userOrderedDataView) {
-          deleteObjectResult = yield* this.deleteObjectInOrderedList(rowId, entity, this.userOrderedDataView);
-        }else{
+        if (targetDataView.orderMember) {
+          deleteObjectResult = yield* this.deleteObjectInOrderedList(rowId, entity, targetDataView);
+        } else {
           deleteObjectResult = yield api.deleteObject({
             SessionFormIdentifier: getSessionId(this),
             Entity: entity,
@@ -813,27 +822,29 @@ export class FormScreenLifecycle02 implements IFormScreenLifecycle02 {
     }
   }
 
-  private *deleteObjectInOrderedList(rowId: string, entity: string, dataView: IDataView) {
+  private *deleteObjectInOrderedList(rowId: string, entity: string, targetDataView: IDataView) {
     const api = getApi(this);
-    const rowToDelete = dataView.dataTable.getRowById(rowId)!;
-    const orderProperty = dataView.orderProperty!;
+    const rowToDelete = targetDataView.dataTable.getRowById(rowId)!;
+    const orderMember = targetDataView.orderMember;
     const newRowOrderMap = {} as any;
-    if (orderProperty) {
-      dataView.dataTable.allRows
-        .filter(row => row[orderProperty.dataIndex] > rowToDelete[orderProperty.dataIndex])
-        .forEach(row => {
-          const rowId = dataView.dataTable.getRowId(row);
-          const newOrder = row[orderProperty.dataIndex] - 1;
+    if (orderMember) {
+      const dataSourceField = getDataSourceFieldByName(targetDataView, orderMember)!;
+      targetDataView.dataTable.allRows
+        .filter((row) => row[dataSourceField.index] > rowToDelete[dataSourceField.index])
+        .forEach((row) => {
+          const rowId = targetDataView.dataTable.getRowId(row);
+          const newOrder = row[dataSourceField.index] - 1;
           newRowOrderMap[rowId] = newOrder;
-        })
+        });
+
+      return yield api.deleteObjectInOrderedList({
+        SessionFormIdentifier: getSessionId(this),
+        Entity: entity,
+        Id: rowId,
+        OrderProperty: orderMember,
+        UpdatedOrderValues: newRowOrderMap,
+      });
     }
-    return yield api.deleteObjectInOrderedList({
-      SessionFormIdentifier: getSessionId(this),
-      Entity: entity,
-      Id: rowId,
-      OrderProperty: orderProperty.name,
-      UpdatedOrderValues: newRowOrderMap
-    });
   }
 
   *saveSession() {
@@ -851,15 +862,17 @@ export class FormScreenLifecycle02 implements IFormScreenLifecycle02 {
         const processQueryInfoResult = yield* processActionQueryInfo(this)(queryResult);
         if (!processQueryInfoResult.canContinue) return;
         result = yield api.saveSession(getSessionId(this));
-        getFormScreen(this).dataViews
-          .forEach(dataView => dataView.dataTable.unlockAddedRowPosition());
+        getFormScreen(this).dataViews.forEach((dataView) =>
+          dataView.dataTable.unlockAddedRowPosition()
+        );
       } finally {
         formScreen.dataUpdateCRS.leave();
       }
       yield* refreshWorkQueues(this)();
       yield* processCRUDResult(this, result);
-      getFormScreen(this).dataViews
-        .forEach(dataView => dataView.dataTable.updateSortAndFilter({retainPreviousSelection: true}));
+      getFormScreen(this).dataViews.forEach((dataView) =>
+        dataView.dataTable.updateSortAndFilter({ retainPreviousSelection: true })
+      );
     } finally {
       this.monitor.inFlow--;
     }
@@ -912,10 +925,10 @@ export class FormScreenLifecycle02 implements IFormScreenLifecycle02 {
     yield* refreshWorkQueues(this)();
   }
 
-  loadInitialData(){
+  loadInitialData() {
     if (!this.isReadData) {
       const self = this;
-      flow(function*(){
+      flow(function* () {
         yield* self.loadData();
       })();
     }
@@ -924,7 +937,7 @@ export class FormScreenLifecycle02 implements IFormScreenLifecycle02 {
   private actionRunning = false;
 
   *executeAction(gridId: string, entity: string, action: IAction, selectedItems: string[]) {
-    if(this.actionRunning){
+    if (this.actionRunning) {
       return;
     }
     this.actionRunning = true;
